[audio] Create audio extension
