--- api/src/main/java/ai/djl/modality/audio/DefaultFeaturizer.java
@@ -1,5 +1,0 @@
-package ai.djl.modality.audio;
-
-public class DefaultFeaturizer implements AudioFeaturizer{
-
-}
