+++ extensions/audio/src/main/java/ai/djl/audio/dataset/AudioData.java
@@ -1,13 +1,10 @@
-package ai.djl.basicdataset.utils;
-
-import ai.djl.ndarray.NDArray;
-import ai.djl.ndarray.NDManager;
-import org.bytedeco.javacv.*;
+package ai.djl.audio.dataset;
 
 import java.nio.Buffer;
 import java.nio.ShortBuffer;
 import java.util.ArrayList;
 import java.util.List;
+import org.bytedeco.javacv.*;
 
 public class AudioData {
     private float[] samples = null;
@@ -25,62 +22,55 @@ public class AudioData {
     public Integer getSampleRate() {
         return sampleRate;
     }
-    public void setSamples(float[] floatArray){
+
+    public void setSamples(float[] floatArray) {
         this.samples = floatArray;
     }
-    public void setSampleRate(Integer rate){
+
+    public void setSampleRate(Integer rate) {
         this.sampleRate = rate;
     }
-    public void setAudioChannels(Integer channels){
+
+    public void setAudioChannels(Integer channels) {
         this.audioChannels = channels;
     }
 
     public AudioData toFloat(String path) throws FrameGrabber.Exception {
         List<Float> floatList = new ArrayList<>();
-        AudioData audioData= new AudioData();
+        AudioData audioData = new AudioData();
         audioData.setAudioChannels(-1);
         audioData.setSampleRate(-1);
-        float scale = (float) 1.0/ (float) (1 << (8 * 2) - 1);
-        try(FFmpegFrameGrabber audioGrabber = new FFmpegFrameGrabber(path)){
+        float scale = (float) 1.0 / (float) (1 << (8 * 2) - 1);
+        try (FFmpegFrameGrabber audioGrabber = new FFmpegFrameGrabber(path)) {
             audioGrabber.start();
             Frame frame;
-            while((frame = audioGrabber.grabFrame()) != null){
+            while ((frame = audioGrabber.grabFrame()) != null) {
                 Buffer[] buffers = frame.samples;
                 ShortBuffer sb = (ShortBuffer) buffers[0];
-                floatList.add (sb.get() * scale);
+                floatList.add(sb.get() * scale);
             }
-        }
-        catch (FrameGrabber.Exception e){
+        } catch (FrameGrabber.Exception e) {
             e.printStackTrace();
         }
-        float [] floatArray = new float[floatList.size()];
-        for (int i = 0; i < floatArray.length ; i++) {
+        float[] floatArray = new float[floatList.size()];
+        for (int i = 0; i < floatArray.length; i++) {
             floatArray[i] = floatList.get(i);
         }
         audioData.setSamples(floatArray);
         return audioData;
     }
 
-
     /**
-     * The configuration for creating a {@link AudioData} value in a {@link
-     *      * ai.djl.training.dataset.Dataset}.
+     * The configuration for creating a {@link AudioData} value in a {@link *
+     * ai.djl.training.dataset.Dataset}.
      */
     public static final class Configuration {
 
-        /**
-         * This parameter is used for setting normalized value.
-         */
+        /** This parameter is used for setting normalized value. */
         private Double target_dB;
-        /**
-         * This parameter is used for setting stride value.
-         */
+        /** This parameter is used for setting stride value. */
         private Double stride_ms;
-        /**
-         * This parameter is used for setting window frame size value.
-         */
+        /** This parameter is used for setting window frame size value. */
         private Double windows_ms;
-
     }
-
 }
