+++ extensions/audio/src/main/java/ai/djl/audio/dataset/SpeechRecognitionDataset.java
@@ -1,13 +1,11 @@
-package ai.djl.basicdataset.Audio;
+package ai.djl.audio.dataset;
 
-import ai.djl.basicdataset.utils.AudioData;
 import ai.djl.basicdataset.utils.TextData;
 import ai.djl.ndarray.NDManager;
 import ai.djl.training.dataset.RandomAccessDataset;
 import ai.djl.training.dataset.Record;
 import ai.djl.translate.TranslateException;
 import ai.djl.util.Progress;
-
 import java.io.IOException;
 
 public abstract class SpeechRecognitionDataset extends RandomAccessDataset {
@@ -16,8 +14,6 @@ public abstract class SpeechRecognitionDataset extends RandomAccessDataset {
     protected TextData recognition;
     protected NDManager manager;
     protected Usage usage;
-
-
 
     /**
      * Creates a new instance of {@link RandomAccessDataset} with the given necessary
@@ -32,9 +28,7 @@ public abstract class SpeechRecognitionDataset extends RandomAccessDataset {
     }
 
     @Override
-    public void prepare(Progress progress) throws IOException, TranslateException {
-
-    }
+    public void prepare(Progress progress) throws IOException, TranslateException {}
 
     @Override
     public Record get(NDManager manager, long index) throws IOException {
@@ -46,7 +40,5 @@ public abstract class SpeechRecognitionDataset extends RandomAccessDataset {
         return 0;
     }
 
-    public abstract static class Builder<T extends Builder<T>> extends BaseBuilder<T>{
-
-    }
+    public abstract static class Builder<T extends Builder<T>> extends BaseBuilder<T> {}
 }
