+++ extensions/audio/src/main/java/ai/djl/audio/featurizer/AudioFeaturizer.java
@@ -1,9 +1,10 @@
-package ai.djl.modality.audio;
+package ai.djl.audio.featurizer;
+
 import ai.djl.ndarray.NDArray;
 import ai.djl.ndarray.NDManager;
 
 public interface AudioFeaturizer {
-    default NDArray ExtractFeatures(NDManager manager, NDArray array) throws Exception{
+    default NDArray ExtractFeatures(NDManager manager, NDArray array) throws Exception {
         return null;
     }
 }
