+++ extensions/audio/src/main/java/ai/djl/audio/featurizer/DefaultFeaturizer.java
@@ -1,0 +1,3 @@
+package ai.djl.audio.featurizer;
+
+public class DefaultFeaturizer implements AudioFeaturizer {}
