+++ settings.gradle
@@ -25,6 +25,7 @@ include ':engines:tensorrt'
 include ':engines:tflite:tflite-engine'
 include ':engines:tflite:tflite-native'
 include ':examples'
+include 'extensions:audio'
 include ':extensions:aws-ai'
 include ':extensions:benchmark'
 include ':extensions:fasttext'
